Move client/util_exec.go to cliutil/exec.go
